+++ src/test/java/org/openlmis/onenetwork/integration/service/OrderableQueueTest.java
@@ -20,7 +20,6 @@ import static org.junit.Assert.assertEquals;
 import java.util.Arrays;
 import java.util.List;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.openlmis.onenetwork.integration.domain.Orderable;
@@ -30,7 +29,10 @@ public class OrderableQueueTest {
   @InjectMocks
   OrderableQueue orderableQueue = new OrderableQueue();
 
-  private Orderable orderable = null;
+  private Orderable orderable = new Orderable.Builder()
+          .withProductCode("testCode")
+          .withFullProductName("testName")
+          .build();
 
   @Test
   public void shouldReturnEmptyList() {
@@ -38,7 +40,6 @@ public class OrderableQueueTest {
   }
 
   @Test
-  @Ignore
   public void shouldAddOrderableToListAndReturnList() {
     orderableQueue.add(orderable);
     List result = orderableQueue.getList();