+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 New functionality added in a backwards-compatible manner:
+* [OLMIS-7304](https://openlmis.atlassian.net/browse/OLMIS-7304): Generating products csv file and sending to sftp server.
 * [OLMIS-7302](https://openlmis.atlassian.net/browse/OLMIS-7302): Created scaffold of one-network-integration-service.
 * [OLMIS-7310](https://openlmis.atlassian.net/browse/OLMIS-7310): Added scheduler.
 * [OLMIS-7309](https://openlmis.atlassian.net/browse/OLMIS-7309): Created connection to SFTP server.